requirements.txt
gliner>=0.2.16
# Optional: only if you set ARMOR_GLINER_USE_MODEL_SPLITTER=1 and the model uses stanza
# stanza>=1.7.0
# langdetect>=1.0.9
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
# Install spacy first (pin to 3.7.x for pre-built wheels; 3.8+ often builds from source and hangs)
spacy>=3.5,<3.8
presidio-analyzer>=2.2.355
flask>=2.3.0
gunicorn>=22.0.0
openai>=1.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
